requirements.txt
# Core Django (Stable)
Django==5.1.4
asgiref==3.8.1
sqlparse==0.5.5
tzdata==2025.3

# Production Server
gunicorn==21.2.0

# Static Files Handling
whitenoise==6.6.0

# Environment Variables
python-dotenv==1.2.2

# Database (Modern psycopg v3 binary)
psycopg[binary]
mysqlclient==2.2.8
dj-database-url==2.1.0

# Authentication
django-allauth==65.15.0

# HTTP Requests
requests==2.32.5

# Google AI / Gemini (Modern v2 SDK)
google-genai==1.67.0

# Voice / TTS Support
gTTS==2.5.4
pyttsx3==2.99

# Security & Utilities
cryptography==46.0.5
PyJWT==2.12.1
typing_extensions==4.15.0
sniffio==1.3.1
tenacity==9.1.4
pydantic==2.12.5
pydantic_core==2.41.5
anyio==4.12.1
httpcore==1.0.9
httpx==0.28.1
certifi==2026.2.25
idna==3.11
urllib3==2.6.3
h11==0.16.0
packaging==26.0
click==8.1.8
distro==1.9.0
typing-inspection==0.4.2
